Move status classes into their own package.
