+++ scos_actions/__init__.py
@@ -1,5 +1,5 @@
-from .status_monitor import StatusMonitor
-from .status_registration_handler import status_registration_handler
+from scos_actions.status.status_monitor import StatusMonitor
+from scos_actions.status.status_registration_handler import status_registration_handler
 from scos_actions.actions.interfaces.signals import register_component_with_status
 
 status_registrar = StatusMonitor()
